Shift around some finals after the merge
