+++ res/style.css
@@ -871,139 +871,6 @@ body, #root, .profileViewer {
   transform: translateY(-30px);
 }
 
-/* Summarize Profile */
-
-.summarize-profile {
-  width: 100%;
-  max-width: 80em;
-  margin: 0 auto;
-  box-sizing: border-box;
-}
-.summarize-profile * {
-  box-sizing: border-box;
-}
-
-.summarize-profile-table {
-  display: flex;
-  flex-direction: column;
-  padding: 2em;
-  /*width: 100%;*/
-  margin: 0 1.66%;
-  font-size: 1.2em;
-}
-
-.summarize-line-graph {
-  min-width: 50%;
-  flex: 1;
-  margin-inline-end: 1em;
-}
-
-.summarize-profile-details {
-  width: 40em;
-  display:flex;
-  align-items: center;
-}
-
-.summarize-profile-numeric {
-  text-align: right;
-  width: 25%;
-}
-
-.summarize-profile-text {
-  flex: 1;
-  flex-shrink: 1;
-}
-
-.summarize-profile-row,
-.summarize-profile-header {
-  display: flex;
-}
-
-.summarize-profile-row:hover {
-  background-color: #fcfcfc;
-}
-
-.summarize-profile-row > div {
-  padding: 0.1em 0;
-}
-
-.summarize-profile-header {
-  text-align: left;
-  font-weight: normal;
-  border-bottom: 1px solid #888;
-  margin-bottom: 0.5em;
-  padding-bottom: 0.5em;
-}
-
-.summarize-profile-numeric {
-  text-align: right;
-  width: 25%;
-}
-
-.summarize-profile-row .summarize-profile-numeric {
-  font-family: monospace;
-  color: #444;
-}
-
-.summarize-profile-thread {
-  margin-bottom: 1em;
-  font-size: 1.5em;
-  color: #000;
-}
-
-.summarize-line-graph svg {
-  background-color:#f6f6f6;
-}
-
-.summarize-profile-filler::after {
-  content: "\00a0";
-}
-
-.summarize-profile-filler {
-  width: 75%;
-  display: inline-block;
-  background-color: #eee;
-}
-
-.summarize-line-graph-filler.filler {
-  background-color: #eee;
-}
-
-.summarize-profile-collapse {
-  background: none;
-  border: none;
-  color: #03f;
-  text-decoration: none;
-  cursor: pointer;
-  padding: 0;
-  margin-left: auto;
-}
-
-.summarize-profile-collapse:hover {
-  text-decoration: underline;
-}
-
-.summarize-profile-collapse::before {
-  content: "";
-  width: 0;
-  height: 0;
-  position: relative;
-  display: inline-block;
-  margin: 0 0.4em;
-  border-left: 0.5em solid transparent;
-  border-right: 0.5em solid transparent;
-  border-top: 0.5em solid #03f;
-  transition: transform 500ms;
-}
-
-.summarize-profile-collapse.expanded::before {
-  transform: rotate(180deg);
-}
-
-.summarize-profile-button-row {
-  text-align: right;
-}
-
 .filler {
   animation-name: filler;
   animation-duration: 1s;
