+++ src/content/components/SummarizeLineGraph.js
